# archiver: Remove cleanup goroutine from BufferPool